added new version for test cases
